requirements.txt
flask
pywebview
pyinstaller
pywin32